+++ spring-whale-database/src/test/java/io/github/springwhale/database/datascope/SmartDataScopeHandlerTest.java
@@ -9,7 +9,6 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.ArgumentCaptor;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
@@ -19,10 +18,7 @@ import java.util.List;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.Mockito.lenient;
-import static org.mockito.Mockito.never;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
+import static org.mockito.Mockito.*;
 
 @ExtendWith(MockitoExtension.class)
 class SmartDataScopeHandlerTest {
@@ -51,7 +47,7 @@ class SmartDataScopeHandlerTest {
         AuthenticationContextHolder.clearContext();
     }
 
-    private void loginAs(Integer userId) {
+    private void loginAs(Long userId) {
         AuthenticationContextHolder.setContext(
                 new AuthenticationContext(userId, "user" + userId, null));
     }
@@ -61,8 +57,8 @@ class SmartDataScopeHandlerTest {
     @Test
     @DisplayName("skipDataScope: cache hit returns cached value without remote call")
     void testSkipDataScopeCacheHit() {
-        loginAs(1);
-        when(cache.get(DataScopeCacheKey.skipDataScope(1), Boolean.class)).thenReturn(true);
+        loginAs(1L);
+        when(cache.get(DataScopeCacheKey.skipDataScope(1L), Boolean.class)).thenReturn(true);
 
         boolean result = handler.skipDataScope();
 
@@ -73,24 +69,24 @@ class SmartDataScopeHandlerTest {
     @Test
     @DisplayName("skipDataScope: cache miss calls remote API and caches result")
     void testSkipDataScopeCacheMiss() {
-        loginAs(1);
-        when(cache.get(DataScopeCacheKey.skipDataScope(1), Boolean.class)).thenReturn(null);
-        when(remoteApi.skipDataScope(1)).thenReturn(new DataScopeSkipResponse(true));
+        loginAs(1L);
+        when(cache.get(DataScopeCacheKey.skipDataScope(1L), Boolean.class)).thenReturn(null);
+        when(remoteApi.skipDataScope(1L)).thenReturn(new DataScopeSkipResponse(true));
 
         boolean result = handler.skipDataScope();
 
         assertThat(result).isTrue();
-        verify(cache).put(DataScopeCacheKey.skipDataScope(1), true, Duration.ofMinutes(5));
-        verify(cache).put(DataScopeCacheKey.fallbackSkipDataScope(1), true, Duration.ofMinutes(30));
+        verify(cache).put(DataScopeCacheKey.skipDataScope(1L), true, Duration.ofMinutes(5));
+        verify(cache).put(DataScopeCacheKey.fallbackSkipDataScope(1L), true, Duration.ofMinutes(30));
     }
 
     @Test
     @DisplayName("skipDataScope: remote API failure falls back to fallback cache")
     void testSkipDataScopeFallbackCacheHit() {
-        loginAs(1);
-        when(cache.get(DataScopeCacheKey.skipDataScope(1), Boolean.class)).thenReturn(null);
-        when(remoteApi.skipDataScope(1)).thenThrow(new RuntimeException("Network error"));
-        when(cache.get(DataScopeCacheKey.fallbackSkipDataScope(1), Boolean.class)).thenReturn(true);
+        loginAs(1L);
+        when(cache.get(DataScopeCacheKey.skipDataScope(1L), Boolean.class)).thenReturn(null);
+        when(remoteApi.skipDataScope(1L)).thenThrow(new RuntimeException("Network error"));
+        when(cache.get(DataScopeCacheKey.fallbackSkipDataScope(1L), Boolean.class)).thenReturn(true);
 
         boolean result = handler.skipDataScope();
 
@@ -100,10 +96,10 @@ class SmartDataScopeHandlerTest {
     @Test
     @DisplayName("skipDataScope: remote API failure with no fallback returns false")
     void testSkipDataScopeFallbackCacheMiss() {
-        loginAs(1);
-        when(cache.get(DataScopeCacheKey.skipDataScope(1), Boolean.class)).thenReturn(null);
-        when(remoteApi.skipDataScope(1)).thenThrow(new RuntimeException("Network error"));
-        when(cache.get(DataScopeCacheKey.fallbackSkipDataScope(1), Boolean.class)).thenReturn(null);
+        loginAs(1L);
+        when(cache.get(DataScopeCacheKey.skipDataScope(1L), Boolean.class)).thenReturn(null);
+        when(remoteApi.skipDataScope(1L)).thenThrow(new RuntimeException("Network error"));
+        when(cache.get(DataScopeCacheKey.fallbackSkipDataScope(1L), Boolean.class)).thenReturn(null);
 
         boolean result = handler.skipDataScope();
 
@@ -122,9 +118,9 @@ class SmartDataScopeHandlerTest {
     @Test
     @DisplayName("skipDataScope: null response from remote API returns false")
     void testSkipDataScopeNullResponse() {
-        loginAs(1);
-        when(cache.get(DataScopeCacheKey.skipDataScope(1), Boolean.class)).thenReturn(null);
-        when(remoteApi.skipDataScope(1)).thenReturn(null);
+        loginAs(1L);
+        when(cache.get(DataScopeCacheKey.skipDataScope(1L), Boolean.class)).thenReturn(null);
+        when(remoteApi.skipDataScope(1L)).thenReturn(null);
 
         boolean result = handler.skipDataScope();
 
@@ -136,8 +132,8 @@ class SmartDataScopeHandlerTest {
     @Test
     @DisplayName("skipTenantScope: cache hit returns cached value without remote call")
     void testSkipTenantScopeCacheHit() {
-        loginAs(1);
-        when(cache.get(DataScopeCacheKey.skipTenantScope(1), Boolean.class)).thenReturn(true);
+        loginAs(1L);
+        when(cache.get(DataScopeCacheKey.skipTenantScope(1L), Boolean.class)).thenReturn(true);
 
         boolean result = handler.skipTenantScope();
 
@@ -148,24 +144,24 @@ class SmartDataScopeHandlerTest {
     @Test
     @DisplayName("skipTenantScope: cache miss calls remote API and caches result")
     void testSkipTenantScopeCacheMiss() {
-        loginAs(1);
-        when(cache.get(DataScopeCacheKey.skipTenantScope(1), Boolean.class)).thenReturn(null);
-        when(remoteApi.skipTenantScope(1)).thenReturn(new DataScopeSkipResponse(true));
+        loginAs(1L);
+        when(cache.get(DataScopeCacheKey.skipTenantScope(1L), Boolean.class)).thenReturn(null);
+        when(remoteApi.skipTenantScope(1L)).thenReturn(new DataScopeSkipResponse(true));
 
         boolean result = handler.skipTenantScope();
 
         assertThat(result).isTrue();
-        verify(cache).put(DataScopeCacheKey.skipTenantScope(1), true, Duration.ofMinutes(5));
-        verify(cache).put(DataScopeCacheKey.fallbackSkipTenantScope(1), true, Duration.ofMinutes(30));
+        verify(cache).put(DataScopeCacheKey.skipTenantScope(1L), true, Duration.ofMinutes(5));
+        verify(cache).put(DataScopeCacheKey.fallbackSkipTenantScope(1L), true, Duration.ofMinutes(30));
     }
 
     @Test
     @DisplayName("skipTenantScope: remote API failure falls back to fallback cache")
     void testSkipTenantScopeFallbackCacheHit() {
-        loginAs(1);
-        when(cache.get(DataScopeCacheKey.skipTenantScope(1), Boolean.class)).thenReturn(null);
-        when(remoteApi.skipTenantScope(1)).thenThrow(new RuntimeException("Network error"));
-        when(cache.get(DataScopeCacheKey.fallbackSkipTenantScope(1), Boolean.class)).thenReturn(true);
+        loginAs(1L);
+        when(cache.get(DataScopeCacheKey.skipTenantScope(1L), Boolean.class)).thenReturn(null);
+        when(remoteApi.skipTenantScope(1L)).thenThrow(new RuntimeException("Network error"));
+        when(cache.get(DataScopeCacheKey.fallbackSkipTenantScope(1L), Boolean.class)).thenReturn(true);
 
         boolean result = handler.skipTenantScope();
 
@@ -175,10 +171,10 @@ class SmartDataScopeHandlerTest {
     @Test
     @DisplayName("skipTenantScope: remote API failure with no fallback returns false")
     void testSkipTenantScopeFallbackCacheMiss() {
-        loginAs(1);
-        when(cache.get(DataScopeCacheKey.skipTenantScope(1), Boolean.class)).thenReturn(null);
-        when(remoteApi.skipTenantScope(1)).thenThrow(new RuntimeException("Network error"));
-        when(cache.get(DataScopeCacheKey.fallbackSkipTenantScope(1), Boolean.class)).thenReturn(null);
+        loginAs(1L);
+        when(cache.get(DataScopeCacheKey.skipTenantScope(1L), Boolean.class)).thenReturn(null);
+        when(remoteApi.skipTenantScope(1L)).thenThrow(new RuntimeException("Network error"));
+        when(cache.get(DataScopeCacheKey.fallbackSkipTenantScope(1L), Boolean.class)).thenReturn(null);
 
         boolean result = handler.skipTenantScope();
 
@@ -199,8 +195,8 @@ class SmartDataScopeHandlerTest {
     @Test
     @DisplayName("resolveDeptIds: cache hit returns cached list without remote call")
     void testResolveDeptIdsCacheHit() {
-        loginAs(1);
-        String cacheKey = DataScopeCacheKey.resolveDeptIds(1, DataScopeType.DEPT, "order");
+        loginAs(1L);
+        String cacheKey = DataScopeCacheKey.resolveDeptIds(1L, DataScopeType.DEPT, "order");
         List<Object> cachedList = List.of(1L, 2L);
         when(cache.getList(cacheKey)).thenReturn(cachedList);
 
@@ -213,31 +209,31 @@ class SmartDataScopeHandlerTest {
     @Test
     @DisplayName("resolveDeptIds: cache miss calls remote API and caches result")
     void testResolveDeptIdsCacheMiss() {
-        loginAs(1);
-        String cacheKey = DataScopeCacheKey.resolveDeptIds(1, DataScopeType.DEPT, "order");
+        loginAs(1L);
+        String cacheKey = DataScopeCacheKey.resolveDeptIds(1L, DataScopeType.DEPT, "order");
         List<Object> deptIds = List.of(1L, 2L, 3L);
         when(cache.getList(cacheKey)).thenReturn(null);
-        when(remoteApi.resolveDeptIds(1, DataScopeType.DEPT, "order"))
+        when(remoteApi.resolveDeptIds(1L, DataScopeType.DEPT, "order"))
                 .thenReturn(new DataScopeResolveResponse(deptIds));
 
         List<Object> result = handler.resolveDeptIds(DataScopeType.DEPT, "order");
 
         assertThat(result).containsExactly(1L, 2L, 3L);
         verify(cache).put(cacheKey, deptIds, Duration.ofMinutes(2));
-        verify(cache).put(DataScopeCacheKey.fallbackResolveDeptIds(1, DataScopeType.DEPT, "order"),
+        verify(cache).put(DataScopeCacheKey.fallbackResolveDeptIds(1L, DataScopeType.DEPT, "order"),
                 deptIds, Duration.ofMinutes(30));
     }
 
     @Test
     @DisplayName("resolveDeptIds: remote API failure falls back to fallback cache")
     void testResolveDeptIdsFallbackCacheHit() {
-        loginAs(1);
-        String cacheKey = DataScopeCacheKey.resolveDeptIds(1, DataScopeType.DEPT, "order");
+        loginAs(1L);
+        String cacheKey = DataScopeCacheKey.resolveDeptIds(1L, DataScopeType.DEPT, "order");
         List<Object> fallbackList = List.of(1L);
         when(cache.getList(cacheKey)).thenReturn(null);
-        when(remoteApi.resolveDeptIds(1, DataScopeType.DEPT, "order"))
+        when(remoteApi.resolveDeptIds(1L, DataScopeType.DEPT, "order"))
                 .thenThrow(new RuntimeException("Network error"));
-        when(cache.getList(DataScopeCacheKey.fallbackResolveDeptIds(1, DataScopeType.DEPT, "order")))
+        when(cache.getList(DataScopeCacheKey.fallbackResolveDeptIds(1L, DataScopeType.DEPT, "order")))
                 .thenReturn(fallbackList);
 
         List<Object> result = handler.resolveDeptIds(DataScopeType.DEPT, "order");
@@ -248,12 +244,12 @@ class SmartDataScopeHandlerTest {
     @Test
     @DisplayName("resolveDeptIds: remote API failure with no fallback returns empty list")
     void testResolveDeptIdsFallbackCacheMiss() {
-        loginAs(1);
-        String cacheKey = DataScopeCacheKey.resolveDeptIds(1, DataScopeType.DEPT, "order");
+        loginAs(1L);
+        String cacheKey = DataScopeCacheKey.resolveDeptIds(1L, DataScopeType.DEPT, "order");
         when(cache.getList(cacheKey)).thenReturn(null);
-        when(remoteApi.resolveDeptIds(1, DataScopeType.DEPT, "order"))
+        when(remoteApi.resolveDeptIds(1L, DataScopeType.DEPT, "order"))
                 .thenThrow(new RuntimeException("Network error"));
-        when(cache.getList(DataScopeCacheKey.fallbackResolveDeptIds(1, DataScopeType.DEPT, "order")))
+        when(cache.getList(DataScopeCacheKey.fallbackResolveDeptIds(1L, DataScopeType.DEPT, "order")))
                 .thenReturn(null);
 
         List<Object> result = handler.resolveDeptIds(DataScopeType.DEPT, "order");
@@ -273,10 +269,10 @@ class SmartDataScopeHandlerTest {
     @Test
     @DisplayName("resolveDeptIds: null response from remote API returns empty list")
     void testResolveDeptIdsNullResponse() {
-        loginAs(1);
-        String cacheKey = DataScopeCacheKey.resolveDeptIds(1, DataScopeType.DEPT, "order");
+        loginAs(1L);
+        String cacheKey = DataScopeCacheKey.resolveDeptIds(1L, DataScopeType.DEPT, "order");
         when(cache.getList(cacheKey)).thenReturn(null);
-        when(remoteApi.resolveDeptIds(1, DataScopeType.DEPT, "order")).thenReturn(null);
+        when(remoteApi.resolveDeptIds(1L, DataScopeType.DEPT, "order")).thenReturn(null);
 
         List<Object> result = handler.resolveDeptIds(DataScopeType.DEPT, "order");
 
@@ -292,18 +288,18 @@ class SmartDataScopeHandlerTest {
         DataScopeProperties.Cache cacheProps = new DataScopeProperties.Cache();
         cacheProps.setSkipTtl(Duration.ofMinutes(10));
         cacheProps.setDeptTtl(Duration.ofMinutes(5));
-        cacheProps.setFallbackTtl(Duration.ofHours(1));
+        cacheProps.setFallbackTtl(Duration.ofHours(1L));
         customProps.setCache(cacheProps);
 
         SmartDataScopeHandler customHandler = new SmartDataScopeHandler(cacheManager, remoteApi, customProps);
 
-        loginAs(1);
-        when(cache.get(DataScopeCacheKey.skipDataScope(1), Boolean.class)).thenReturn(null);
-        when(remoteApi.skipDataScope(1)).thenReturn(new DataScopeSkipResponse(true));
+        loginAs(1L);
+        when(cache.get(DataScopeCacheKey.skipDataScope(1L), Boolean.class)).thenReturn(null);
+        when(remoteApi.skipDataScope(1L)).thenReturn(new DataScopeSkipResponse(true));
 
         customHandler.skipDataScope();
 
-        verify(cache).put(eq(DataScopeCacheKey.skipDataScope(1)), eq(true), eq(Duration.ofMinutes(10)));
-        verify(cache).put(eq(DataScopeCacheKey.fallbackSkipDataScope(1)), eq(true), eq(Duration.ofHours(1)));
+        verify(cache).put(eq(DataScopeCacheKey.skipDataScope(1L)), eq(true), eq(Duration.ofMinutes(10)));
+        verify(cache).put(eq(DataScopeCacheKey.fallbackSkipDataScope(1L)), eq(true), eq(Duration.ofHours(1L)));
     }
 }
