+++ spring-whale-platform/rbac/spring-whale-platform-rbac/src/main/java/io/github/springwhale/platform/rbac/dao/entity/RoleDeptEntity.java
@@ -14,8 +14,8 @@ import lombok.EqualsAndHashCode;
 public class RoleDeptEntity extends BaseEntity {
 
     @Column(nullable = false)
-    private Integer roleId;
+    private Long roleId;
 
     @Column(nullable = false)
-    private Integer groupId;
+    private Long groupId;
 }
